Move org dashboards to use react router
